refactor(moe): move moe_forward_*_impl to moe/forward.rs (Phase 3a)
The 3 free-function MoE dispatchers (`moe_forward_stacked_decode_impl`,
`moe_forward_batched_prefill_impl`, `moe_forward_batched_decode_impl`)
were declared at the bottom of qwen3_moe.rs (lines 2912-3539) — 600+
lines below the model that calls them. Per the 2026-05-09 audit,
qwen3_moe.rs at 3579 lines was a top size offender, and the MoE
dispatch was a half-done extraction (`moe/dispatch.rs` had a `moe_forward<B>`
placeholder but production used these inline impls).

This commit relocates them to `crates/ferrum-models/src/moe/forward.rs`
without changing semantics:

  qwen3_moe.rs:  3579 → 2950 lines  (-629)
  moe/forward.rs: created at 674 lines

Type signatures unchanged: still take `Qwen3MoeLayerState<B>` +
`Qwen3MoeScratch<B>` by reference + scratch by mut ref (so the caller
can split the borrow on `Qwen3MoeModel`). Profiling counters
(DEC_ROUTE_US, MOE_PREFILL_*, etc.) made `pub(crate)` and imported.

Call sites in qwen3_moe.rs (4 of them) updated to
`crate::moe::forward::moe_forward_*_impl::<B>(...)`.

Acceptance gates:
  cargo check --workspace --all-targets  clean
  cargo fmt --all -- --check             clean
  cargo test --workspace --lib           all green
  RUSTFLAGS=-D warnings cargo check      clean
  Metal smoke (Qwen3-0.6B, 3×128 tok):
    Phase 2.5 baseline: 60.5 tok/s
    Phase 3a:           61.2 tok/s (+1.2%, within ±3%)
  Metal MoE quick smoke (Qwen3-30B-A3B Q4_K_M, 1×32):
    decode 43.0 tok/s — moved fns exercise normally

This is the first chunk of Phase 3 (Linear/MoE-trait decoupling). The
`B::*quant*` calls inside the moved fns stay; the next sub-phase will
move them into ExpertStack methods so model code stops calling backend
quant methods directly.
